requirements.txt
absl-py==2.1.0
astunparse==1.6.3
bcrypt==4.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
deepface==0.0.93
filelock==3.15.4
fire==0.6.0
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Login==0.6.3
flatbuffers==24.3.25
fsspec==2024.6.1
gast==0.6.0
gdown==5.2.0
google-pasta==0.2.0
grpcio==1.66.1
gunicorn==23.0.0
h5py==3.11.0
huggingface-hub==0.24.6
humanfriendly==10.0
idna==3.8
imgbeddings==0.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
keras==3.5.0
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
mtcnn==0.1.1
namex==0.0.8
networkx==3.3
numpy==1.26.4
onnxruntime==1.19.0
opencv-python==4.10.0.84
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pandas==2.2.2
pillow==10.4.0
protobuf==4.25.4
psycopg2-binary==2.9.9
Pygments==2.18.0
pyreadline3==3.4.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
retina-face==0.0.17
rich==13.8.1
safetensors==0.4.4
scikit-learn==1.5.1
scipy==1.14.1
setuptools==75.0.0
six==1.16.0
soupsieve==2.6
sympy==1.13.2
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-intel==2.17.0
termcolor==2.4.0
tf_keras==2.17.0
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.4.1
tqdm==4.66.5
transformers==4.44.2
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
waitress==3.0.0
Werkzeug==3.0.4
wheel==0.44.0
wrapt==1.16.0
